# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resume Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/experiences/write": {
      "post": {
        "summary": "Write Experience",
        "operationId": "write_experience_api_experiences_write_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__experience__api__WritingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiences/draft": {
      "post": {
        "summary": "Draft Experience",
        "operationId": "draft_experience_api_experiences_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritingResult"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/experiences/metadata": {
      "post": {
        "summary": "Generate Metadata",
        "operationId": "generate_metadata_api_experiences_metadata_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetadataInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataResult"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DraftInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title"
          },
          "period": {
            "type": "string",
            "maxLength": 200,
            "title": "Period"
          },
          "markdown": {
            "type": "string",
            "maxLength": 60000,
            "title": "Markdown",
            "default": ""
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Sources"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "period"
        ],
        "title": "DraftInput"
      },
...
      "ExperienceInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title"
          },
          "period": {
            "type": "string",
            "maxLength": 200,
            "title": "Period"
          },
          "markdown": {
            "type": "string",
            "maxLength": 60000,
            "title": "Markdown",
            "default": ""
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Sources"
          },
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "metadata": {
            "type": "string",
            "maxLength": 12000,
            "title": "Metadata",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "period",
          "id",
          "revision"
        ],
        "title": "ExperienceInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialInput": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "title": "Content"
          },
          "materialType": {
            "type": "string",
            "enum": [
              "company",
              "jobPosting"
            ],
            "title": "Materialtype"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "content",
          "materialType"
        ],
        "title": "MaterialInput"
      },
...
      "MetadataInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "title": "Title"
          },
          "period": {
            "type": "string",
            "maxLength": 200,
            "title": "Period"
          },
          "markdown": {
            "type": "string",
            "maxLength": 60000,
            "minLength": 1,
            "title": "Markdown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "period",
          "markdown"
        ],
        "title": "MetadataInput"
      },
      "MetadataResult": {
        "properties": {
          "metadata": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metadata"
        ],
        "title": "MetadataResult"
      },
...
      "ResumeInput": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "text": {
            "type": "string",
            "maxLength": 100000,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text"
        ],
        "title": "ResumeInput"
      },
...
      "Source": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "note",
              "link",
              "file"
            ],
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "maxLength": 500,
            "title": "Name"
          },
          "text": {
            "type": "string",
            "maxLength": 100000,
            "title": "Text"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "createdAt": {
            "type": "string",
            "maxLength": 100,
            "title": "Createdat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "name",
          "text",
          "createdAt"
        ],
        "title": "Source"
      },
...
      "SourceNote-Output": {
        "properties": {
          "sourceId": {
            "type": "string",
            "title": "Sourceid"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "sourceId",
          "text",
          "verified"
        ],
        "title": "SourceNote"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WritingResult": {
        "properties": {
          "markdown": {
            "type": "string",
            "maxLength": 60000,
            "minLength": 1,
            "title": "Markdown"
          },
          "summary": {
            "type": "string",
            "maxLength": 10000,
            "title": "Summary"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Questions"
          },
          "sourceNotes": {
            "items": {
              "$ref": "#/components/schemas/SourceNote-Output"
            },
            "type": "array",
            "title": "Sourcenotes"
          }
        },
        "type": "object",
        "required": [
          "markdown",
          "summary",
          "questions",
          "sourceNotes"
        ],
        "title": "WritingResult"
      },
...
      "app__experience__api__WritingInput": {
        "properties": {
          "experience": {
            "$ref": "#/components/schemas/ExperienceInput"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeInput"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialInput"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Materials"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "experience",
          "resume",
          "materials"
        ],
        "title": "WritingInput"
      },
...
    }
  }
}